+++ tools/osdtest.c
@@ -28,9 +28,9 @@ void unmap_regspace()
 	close(regfd);
 }
 
-void write_register(long reg, unsigned long value)
+void write_register(__u32 reg, __u32 value)
 {
-	unsigned long regval[2] = {reg, value};
+	__u32 regval[2] = {reg, value};
 	if(!regfd)
 		return;
 	
@@ -39,7 +39,7 @@ void write_register(long reg, unsigned long value)
 	}
 }
 
-unsigned long read_register(long reg)
+__u32 read_register(__u32 reg)
 {
 	if(ioctl(regfd, TL880IOCREADREG, &reg) < 0) {
 		perror("Unable to read register");
@@ -49,7 +49,7 @@ unsigned long read_register(long reg)
 }
 
 
-unsigned long read_regbits(long reg, long high_bit, long low_bit)
+__u32 read_regbits(__u32 reg, int high_bit, int low_bit)
 {
 	int shift = high_bit - low_bit + 1;
 	int mask = ~(0xffffffff << shift);
@@ -57,9 +57,9 @@ unsigned long read_regbits(long reg, long high_bit, long low_bit)
 	return (read_register(reg) >> low_bit) & mask;
 }
 
-void set_bits(unsigned long *value, long reg, long high_bit, long low_bit, unsigned long setvalue)
+void set_bits(__u32 *value, __u32 reg, int high_bit, int low_bit, __u32 setvalue)
 {
-	register unsigned long mask = 0;
+	register __u32 mask = 0;
 
 	/* set bits from high_bit to low_bit in mask to 1 */
 	mask = ~(0xFFFFFFFF << (high_bit - low_bit + 1)) << low_bit;
@@ -72,9 +72,9 @@ void set_bits(unsigned long *value, long reg, long high_bit, long low_bit, unsig
 
 
 
-void write_regbits(long reg, long high_bit, long low_bit, unsigned long value)
+void write_regbits(__u32 reg, int high_bit, int low_bit, __u32 value)
 {
-	unsigned long curval = read_register(reg);
+	__u32 curval = read_register(reg);
 
 	set_bits(&curval, reg, high_bit, low_bit, value);
 
@@ -88,7 +88,7 @@ void write_regbits(long reg, long high_bit, long low_bit, unsigned long value)
  * Range: 0-255 all channels
  * Return: value suitable for writing to regs 10140-1017c (cursor palette)
  */
-unsigned long rgb_to_ypbpr(unsigned long r, unsigned long g, unsigned long b, unsigned long a)
+__u32 rgb_to_ypbpr(__u8 r, __u8 g, __u8 b, __u8 a)
 {
 	float red = (float)r / 255.0;
 	float green = (float)g / 255.0;
@@ -131,11 +131,11 @@ unsigned long rgb_to_ypbpr(unsigned long r, unsigned long g, unsigned long b, un
 
 	/* Use r, g, b, a as temporary storage variables to save namespace */
 	/* New Y */
-	r = (unsigned long)(y * 255.0);
+	r = (__u32)(y * 255.0);
 	/* New Pb */
-	g = (unsigned long)(pb * 255.0);
+	g = (__u32)(pb * 255.0);
 	/* New Pr */
-	b = (unsigned long)(pr * 255.0);
+	b = (__u32)(pr * 255.0);
 	/* New A */
 	a = alpha * 127.0;
 
@@ -148,7 +148,7 @@ int main(int argc, char *argv[])
 {
 	int memfd, i, j;
 	unsigned char *memspace;
-	unsigned long *lmspace;
+	__u32 *lmspace;
 	unsigned char r1 = 0xff;
 	unsigned char g1 = 0x00;
 	unsigned char b1 = 0x00;
@@ -197,7 +197,7 @@ int main(int argc, char *argv[])
 		unmap_regspace();
 		return -1;
 	}
-	lmspace = (unsigned long *)memspace;
+	lmspace = (__u32 *)memspace;
 
 	/*
 	for(i = 0x10088; i <= 0x100ac; i += 4) {
